+++ runtime/models/models.go
@@ -16,15 +16,12 @@ import (
 	"strings"
 
 	"github.com/hypermodeinc/modus/lib/manifest"
-	"github.com/hypermodeinc/modus/runtime/config"
 	"github.com/hypermodeinc/modus/runtime/db"
 	"github.com/hypermodeinc/modus/runtime/hosts"
 	"github.com/hypermodeinc/modus/runtime/manifestdata"
 	"github.com/hypermodeinc/modus/runtime/secrets"
 	"github.com/hypermodeinc/modus/runtime/utils"
 )
-
-var localHypermodeModels = map[string]bool{"meta-llama/meta-llama-3.1-8b-instruct": true, "sentence-transformers/all-minilm-l6-v2": true, "antoinemc/distilbart-mnli-github-issues": true, "distilbert/distilbert-base-uncased-finetuned-sst-2-english": true}
 
 func GetModel(modelName string) (*manifest.ModelInfo, error) {
 	model, ok := manifestdata.GetManifest().Models[modelName]
@@ -67,7 +64,13 @@ func PostToModelEndpoint[TResult any](ctx context.Context, model *manifest.Model
 	span, ctx := utils.NewSentrySpanForCurrentFunc(ctx)
 	defer span.Finish()
 
-	endpoint, host, err := getModelEndpointAndHost(model)
+	host, err := hosts.GetHttpHost(model.Host)
+	if err != nil {
+		var empty TResult
+		return empty, err
+	}
+
+	url, err := getModelEndpointUrl(model, host)
 	if err != nil {
 		var empty TResult
 		return empty, err
@@ -75,16 +78,14 @@ func PostToModelEndpoint[TResult any](ctx context.Context, model *manifest.Model
 
 	bs := func(ctx context.Context, req *http.Request) error {
 		req.Header.Set("Content-Type", "application/json")
-		if host.Name != hosts.HypermodeHost {
+		if host.Name == hosts.HypermodeHost {
+			return authenticateHypermodeModelRequest(ctx, req, host)
+		} else {
 			return secrets.ApplyHostSecretsToHttpRequest(ctx, host, req)
 		}
-		if config.IsDevEnvironment() {
-			return secrets.ApplyAuthToLocalModelRequest(ctx, host, req)
-		}
-		return nil
 	}
 
-	res, err := utils.PostHttp[TResult](ctx, endpoint, payload, bs)
+	res, err := utils.PostHttp[TResult](ctx, url, payload, bs)
 	if err != nil {
 		var empty TResult
 		return empty, err
@@ -95,47 +96,27 @@ func PostToModelEndpoint[TResult any](ctx context.Context, model *manifest.Model
 	return res.Data, nil
 }
 
-func getModelEndpointAndHost(model *manifest.ModelInfo) (string, *manifest.HTTPHostInfo, error) {
-
-	host, err := hosts.GetHttpHost(model.Host)
-	if err != nil {
-		return "", nil, err
-	}
+func getModelEndpointUrl(model *manifest.ModelInfo, host *manifest.HTTPHostInfo) (string, error) {
 
 	if host.Name == hosts.HypermodeHost {
-		if config.IsDevEnvironment() {
-			if !isValidLocalHypermodeModel(model.SourceModel) {
-				return "", host, fmt.Errorf("model %s is not available in the local Hypermode environment", model.SourceModel)
-			}
-			endpoint := fmt.Sprintf("https://models.hypermode.host/%s", strings.ToLower(model.SourceModel))
-			return endpoint, host, nil
-		}
-		// Compose the Hypermode hosted model endpoint URL.
-		// Example: http://modelname.bckid.svc.cluster.local/v1/models/modelname:predict
-		endpoint := fmt.Sprintf("http://%s.%s/%[1]s:predict", strings.ToLower(model.Name), config.ModelHost)
-		return endpoint, host, nil
+		return getHypermodeModelEndpointUrl(model)
 	}
 
 	if host.BaseURL != "" && host.Endpoint != "" {
-		return "", host, fmt.Errorf("specify either base URL or endpoint for a host, not both")
+		return "", fmt.Errorf("specify either base URL or endpoint for a host, not both")
 	}
 
 	if host.BaseURL != "" {
 		if model.Path == "" {
-			return "", host, fmt.Errorf("model path is not defined")
+			return "", fmt.Errorf("model path is not defined")
 		}
 		endpoint := fmt.Sprintf("%s/%s", strings.TrimRight(host.BaseURL, "/"), strings.TrimLeft(model.Path, "/"))
-		return endpoint, host, nil
+		return endpoint, nil
 	}
 
 	if model.Path != "" {
-		return "", host, fmt.Errorf("model path is defined but host has no base URL")
+		return "", fmt.Errorf("model path is defined but host has no base URL")
 	}
 
-	return host.Endpoint, host, nil
-}
-
-func isValidLocalHypermodeModel(modelName string) bool {
-	_, ok := localHypermodeModels[strings.ToLower(modelName)]
-	return ok
+	return host.Endpoint, nil
 }