+++ CHANGELOG.md
@@ -33,6 +33,7 @@ In previous releases, the name "Hypermode" was used for all three._
 - Remove AWS Secrets Manager client [#456](https://github.com/hypermodeinc/modus/pull/456)
 - Make app path required [#457](https://github.com/hypermodeinc/modus/pull/457)
 - Improve `.env` file handling [#458](https://github.com/hypermodeinc/modus/pull/458)
+- Update command-line args and env variables [#459](https://github.com/hypermodeinc/modus/pull/459)
 
 ## 2024-10-02 - Version 0.12.7